Document what exactly Matrix4f.translate() does.
Matrix4f.translate() was unintuitive, since it merely altered the first
three components of the last column. But it was used as though one can
achieve a translation to an already existing transformation, which is
not the case!

Now, it merely builds a complete new matrix which only represents a
translation and must be multiplied against another transformation
matrix.

Also, add zero() methods to vectors.