Rename the float/double precision constant.
Reason being that the precision cannot be expressed in terms of a
constant number, but must rather be expressed as ULP with respect to a
given magnitude.
Therefore, that approximate ULP is with respect to small values in
magnitude (1.0 or 25.0 compared to 1.0E20 or 25.0E32). Make that clear
in the constant names.