Add Quaternion.lookAlong
This method creates a rotation that maps the given direction to the
positive Z axis.
This is unlike Matrix4f.lookAlong, which does what GLU's lookAt does,
namely mapping the direction to the negative Z axis! We need to figure
whether we want this or make both classes use the same semantics.