Change Quaternion.lookAlong to .rotateDirToPositiveZ
This method now applies a transformation to an existing quaternion that
rotates the 'dir' to +Z. Also renamed this because it behaves
differently compared to lookAt/lookAlong.
But why do we need different behaviour? This allows to chain multiple
rotation transformations that do not swap Z all the time. You usually
would not chain multiple GLU lookAt together, but with this new
quaternion method, you can. :)