Add Matrix4.isAabInsideFrustumMasked()
Allows to mask-off a set of frustum planes that should not be tested
against the box by assuming that the box is on the positive side of that
plane.
This can be used to accelerate frustum checks where one knows that the
box is likely to reside only on a certain side of the frustum, or to
disable checking against the far frustum plane.