Remove the "fast path" in Matrix.normal()
Determining whether a matrix is orthogonal is more expensive than just
computing the transpose of the inverse.
The user needs to know whether a matrix is possibly an orthogonal matrix
and then not use that normal() method but set3x3() instead to obtain the
rotation part.

Also: improve determinant computation