Simplify the MatrixStack classes by subclassing Matrix4
Thanks for @theagentd for this cool suggestion!

This means that we now have to define some behaviour regarding
subtyping:
- Matrix4.equals/hashCode: MatrixStack does not overwrite those methods.
Additionally, the implementation of Matrix4.equals was changed to permit
subtypes of Matrix4 to be equal to Matrix4 instances. This also means
that if someone stores a MatrixStack and a Matrix4 in a hashed
collection, both Matrix4 and MatrixStack will be treated as equal
objects iff the values of the Matrix4 and the "current" matrix in the
MatrixStack are equal!
- operations on Matrix4 that will be given a MatrixStack as parameter
effectively will be given the "current" matrix in that MatrixStack