Add bounding rectangle test and add interval tree
The interval tree helps to improve performance _dramatically_ by not
needing to test all polygon edges anymore but only those that actually
intersect the imaginary ray along the y co-ordinate of the search point.
With this method, polygons with more than 64,000 vertices can be tested
against around 30 million points in around a second!

I searched for an algorithm that does this with an interval tree, but
could not find one so far. It's an own implementation.