Add PolygonsPointIntersection to test many polygons at once
This class is a generalization of PolygonPointIntersection where all the
edges of all polygons are inserted as intervals into the interval tree
and annotated with the polygon they belong to.
During querying a bitset is used to maintain even/odd states per polygon
and the bitset is returned as result, which will have it's i-th bit set
iff the query point lies inside the i-th polygon.