Add Quaternion.integrate()
This adds a scaled rotation given as a elapsed time delta and the
angular velocities around local coordinate axes of the rotation it
should be applied to.
It is very handy to update the rotation/orientation of an object based
on angular velocity (resulting from angular accceleration/torque for
example) and elapsed time.