Add optimizations by forbidding direct Matrix4f field accesses
We keep track of the "properties" of a matrix in the modifying methods
to select the optimal method at runtime. For example, when the matrix is
the identity, any invocation to one of the post-multiplying methods
translate/rotate/scale/lookAt/perspective/ortho/... can be routed to a
non-multiplying method invocation to save the unnecessary
multiplications.