Improve performance of Matrix.invert() for affine and orthonormal case
In the very common case of orthonormal, affine matrices, built from
rotate and translate and their derivates such as lookAt/lookAlong, etc.,
we can improve the performance of invert 2x. Since we compute the 3x3
determinant anyway to perform generic inversion, we can check the
determinant for orthonormality of the matrix.