Preserve properties as best as possible when manually setting elements
Some people tend to use JOML by not using its provided methods (like
translate, rotate, scale, lookAt, perspective, etc.) but writing those
methods themselves and setting the individual matrix elements.

This change is to preserve matrix properties as best as possible for
further operations to be as fast as possible.