Rewrite invokedynamic for string concatenation to StringBuffer
Since JEP 280 javac generates invokedynamic instructions for string
concatenations when targeting Java 9+.
We need target = 9 for compiling against Panama Vector API because
we need to be able to enable Java 9 modules during compilation. But we
still want to be able to backport the bytecode to earlier Java versions
(currently Java 1.2).